#93 [Feature] 일정 설정 완료 화면 & 장소 검색 화면 마커 표시

+++ app/src/main/java/com/thequietz/travelog/schedule/model/SchedulePlaceModel.kt
@@ -13,10 +13,10 @@ data class SchedulePlaceModel(
     val areaCode: Int,
 
     @SerializedName("mapx")
-    val mapX: Float,
+    val mapX: Double,
 
     @SerializedName("mapy")
-    val mapY: Float,
+    val mapY: Double,
 
     @SerializedName("cityname")
     val cityName: String,
